fix: separate types with comments
